requirements.txt
alembic==1.7.1
asyncpg==0.24.0
fastapi==0.68.1
sqlmodel==0.0.4
uvicorn==0.15.0
PyJWT
passlib
python-dotenv
pydantic[email]